auth-payment/FirebaseFunctions/requirements.txt
firebase_functions~=0.1.0
stripe
python-dotenv
